refactor: 状態ごとに切り替えるコンポーネントを分割
