+++ src/app/ar-assets/create/_components/types.ts
@@ -1,4 +1,4 @@
-const steps = [0, 1, 2, 3] as const;
+export const steps = [0, 1, 2, 3] as const;
 export type Step = (typeof steps)[number];
 
 type StepInput<T extends Step> = T extends 0
